pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.0",
    "langchain-huggingface>=0.1.0",
    "langchain-openai>=1.1.4",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=3.0.0",
]
